# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zed CV API
  description: AI-powered job matching platform for Zambia
  version: 0.1.0
  contact:
    email: convergeozambia@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.zedcv.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Job:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description: { type: string }
        requirements: { type: array, items: { type: string } }
        skills_required: { type: array, items: { type: string } }
        salary_min: { type: integer }
        salary_max: { type: integer }
        apply_url: { type: string, format: uri }
        apply_email: { type: string, format: email }
        source: { type: string, enum: [manual, scraper, ocr, partner] }
        quality_score: { type: integer, minimum: 0, maximum: 100 }
        closing_date: { type: string, format: date }
        posted_at: { type: string, format: date-time }
        is_active: { type: boolean }
    JobCreate:
      type: object
      required: [title, description, source]
      properties:
        title: { type: string, minLength: 5 }
        company: { type: string }
        location: { type: string }
        description: { type: string, minLength: 20 }
        requirements: { type: array, items: { type: string } }
        skills_required: { type: array, items: { type: string } }
        salary_min: { type: integer }
        salary_max: { type: integer }
        apply_url: { type: string, format: uri }
        apply_email: { type: string, format: email }
        source: { type: string, enum: [manual, scraper, ocr, partner] }
        closing_date: { type: string, format: date }
    JobList:
      type: object
      properties:
        jobs: { type: array, items: { $ref: '#/components/schemas/Job' } }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 20 } }
        - { name: location, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobList' }
    post:
      operationId: createJob
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
...